refactor: создать дао лайков, перенести методы
