+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/FilmDbStorage.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.interfaces.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.interfaces.GenreStorage;
@@ -84,23 +83,6 @@ public class FilmDbStorage implements FilmStorage {
         String sql = "SELECT film_id, title, description, release_date, duration, mpa_id " +
                 "FROM film";
         return jdbcTemplate.query(sql, this::mapRowToFilm);
-    }
-
-    @Override
-    public void addLike(int filmId, int userId) {
-        String sql = "INSERT INTO user_like_film (film_id, user_id) " +
-                "VALUES (?, ?)";
-        jdbcTemplate.update(sql,
-                filmId,
-                userId);
-    }
-
-    @Override
-    public void removeLike(int filmId, int userId) {
-        String sql = "DELETE FROM user_like_film WHERE film_id = ? AND user_id = ?";
-        if (jdbcTemplate.update(sql, filmId, userId) == 0) {
-            throw new NotFoundException("Лайк пользователя c id=" + userId + " фильму c id=" + filmId + " не найден!");
-        }
     }
 
     @Override
