+++ src/main/java/ru/yandex/practicum/filmorate/storage/interfaces/UserLikeFilmStorage.java
@@ -1,0 +1,6 @@
+package ru.yandex.practicum.filmorate.storage.interfaces;
+
+public interface UserLikeFilmStorage {
+    void addLike(int filmId, int userId);
+    void removeLike(int filmId, int userId);
+}
